+++ lib/game/game.dart
@@ -233,7 +233,15 @@ abstract class BaseGame extends Game with TapDetector {
     _addLater.clear();
 
     components.forEach((c) => c.update(t));
-    components.removeWhere((c) => c.destroy());
+    components.removeWhere((c) {
+      final destroy = c.destroy();
+
+      if (destroy) {
+        c.onDestroy();
+      }
+
+      return destroy;
+    });
   }
 
   /// This implementation of resize passes the resize call along to every component in the list, enabling each one to make their decisions as how to handle the resize.